Use a double for damage related keys. (#2174)
Using an int is an impl detail, damage is represented as doubles.

+++ src/main/java/org/spongepowered/api/entity/FallingBlock.java
@@ -54,7 +54,7 @@ public interface FallingBlock extends Entity {
      * {@link Keys#MAX_FALL_DAMAGE}
      * @return The maximum fall damage
      */
-    default Value.Mutable<Integer> maxFallDamage() {
+    default Value.Mutable<Double> maxFallDamage() {
         return this.requireValue(Keys.MAX_FALL_DAMAGE).asMutable();
     }
